requirements.txt
requests==2.28.1
schedule==1.1.0
python-whois==0.7.0
